requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
PyPDF2==3.0.1
python-dotenv==1.0.1
openai>=1.3.0
pydantic==2.6.1
httpx>=0.24.0 